requirements.txt
python-dotenv==1.0.1
qdrant-client==1.12.2
langchain-openai==0.2.14
langchain-community==0.3.13
pandas==2.2.3
langchain_qdrant==0.2.0
